refactor(commands): split order.go into focused files
